+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -12,6 +12,8 @@ group: Deployment & Configs
 
 ### RC 
 
+- Fix on 'Add Parameter' when definition parameter group is not a built-in one.
+
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}